refactor(core): centralize handler execution in HandlerExecutor
- Create IHandlerExecutor interface to abstract handler execution
- Implement HandlerExecutor with centralized reflection and caching
- Move all handler lookup, delegate caching, and invocation logic from
  Mediator to HandlerExecutor
- Refactor Mediator to delegate all handler execution to HandlerExecutor
- Register HandlerExecutor as singleton in DI container
- Update Mediator tests to use IHandlerExecutor instead of IServiceProvider
- Update cache helper methods to access HandlerExecutor instead of Mediator

This change eliminates scattered reflection throughout the codebase by
centralizing all handler execution logic in a single component.
