+++ src/Core/Implementation/Mediator.cs
@@ -1,53 +1,27 @@
 using System;
-using System.Collections.Concurrent;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 using Coordix.Interfaces;
-using Microsoft.Extensions.DependencyInjection;
 
 namespace Coordix.Implementation
 {
 	/// <summary>
 	/// A mediator that resolves handlers from an IServiceProvider and invokes them for requests and notifications.
-	/// This implementation uses performance optimizations including cached MethodInfo and compiled delegates
-	/// to minimize reflection overhead and provide near-native invocation performance.
+	/// This implementation delegates handler execution to IHandlerExecutor, which centralizes all reflection
+	/// and caching logic, eliminating scattered reflection throughout the codebase.
 	/// </summary>
 	public class Mediator : IMediator
 	{
-		private readonly IServiceProvider _provider;
+		private readonly IHandlerExecutor _handlerExecutor;
 
 		/// <summary>
-		/// Cache for MethodInfo objects to avoid repeated reflection calls.
-		/// Key: Handler type, Value: Handle method's MethodInfo.
+		/// Initializes a new instance of the <see cref="Mediator"/> class using the given handler executor.
 		/// </summary>
-		private static readonly ConcurrentDictionary<Type, MethodInfo> _methodInfoCache = new ConcurrentDictionary<Type, MethodInfo>();
-
-		/// <summary>
-		/// Cache for compiled delegates for request handlers without return value.
-		/// Key: Handler type, Value: Compiled delegate for invoking the handler.
-		/// </summary>
-		private static readonly ConcurrentDictionary<Type, Func<object, IRequest, CancellationToken, Task>> _requestHandlerDelegates = new ConcurrentDictionary<Type, Func<object, IRequest, CancellationToken, Task>>();
-
-		/// <summary>
-		/// Cache for compiled delegates for request handlers with return value.
-		/// Key: Handler type, Value: Compiled delegate for invoking the handler.
-		/// </summary>
-		private static readonly ConcurrentDictionary<Type, Delegate> _requestHandlerWithResponseDelegates = new ConcurrentDictionary<Type, Delegate>();
-
-		/// <summary>
-		/// Cache for compiled delegates for notification handlers.
-		/// Key: Handler type, Value: Compiled delegate for invoking the handler.
-		/// </summary>
-		private static readonly ConcurrentDictionary<Type, Func<object, INotification, CancellationToken, Task>> _notificationHandlerDelegates = new ConcurrentDictionary<Type, Func<object, INotification, CancellationToken, Task>>();
-
-		/// <summary>
-		/// Initializes a new instance of the <see cref="Mediator"/> class using the given service provider.
-		/// </summary>
-		/// <param name="provider">The service provider used to resolve handler instances.</param>
-		public Mediator(IServiceProvider provider) => _provider = provider;
+		/// <param name="handlerExecutor">The handler executor used to execute handlers.</param>
+		public Mediator(IHandlerExecutor handlerExecutor)
+		{
+			_handlerExecutor = handlerExecutor ?? throw new ArgumentNullException(nameof(handlerExecutor));
+		}
 
 		/// <summary>
 		/// Sends a request to a single handler and returns its response.
@@ -57,21 +31,9 @@ namespace Coordix.Implementation
 		/// <param name="cancellationToken">A token to monitor for cancellation requests.</param>
 		/// <returns>A task representing the asynchronous operation, containing the handler's response.</returns>
 		/// <exception cref="InvalidOperationException">Thrown if no handler is found for the request type.</exception>
-		public async Task<TResponse> Send<TResponse>(IRequest<TResponse> request, CancellationToken cancellationToken = default)
+		public Task<TResponse> Send<TResponse>(IRequest<TResponse> request, CancellationToken cancellationToken = default)
 		{
-			var requestType = request.GetType();
-			var responseType = typeof(TResponse);
-			var handlerType = typeof(IRequestHandler<,>).MakeGenericType(requestType, responseType);
-
-			var handler = _provider.GetService(handlerType);
-
-			if (handler is null)
-			{
-				throw new InvalidOperationException($"Handler not found for {requestType.Name}");
-			}
-
-			var response = await ToHandleInvokeRequest(handlerType, handler, request, cancellationToken);
-			return response;
+			return _handlerExecutor.ExecuteRequestHandler(request, cancellationToken);
 		}
 
 		/// <summary>
@@ -81,19 +43,9 @@ namespace Coordix.Implementation
 		/// <param name="cancellationToken">A token to monitor for cancellation requests.</param>
 		/// <returns>A task representing the asynchronous operation.</returns>
 		/// <exception cref="InvalidOperationException">Thrown if no handler is found for the request type.</exception>
-		public async Task Send(IRequest request, CancellationToken cancellationToken = default)
+		public Task Send(IRequest request, CancellationToken cancellationToken = default)
 		{
-			var requestType = request.GetType();
-			var handlerType = typeof(IRequestHandler<>).MakeGenericType(requestType);
-
-			var handler = _provider.GetService(handlerType);
-
-			if (handler is null)
-			{
-				throw new InvalidOperationException($"Handler not found for {requestType.Name}");
-			}
-
-			await ToHandleInvokeRequest(handlerType, handler, request, cancellationToken);
+			return _handlerExecutor.ExecuteRequestHandler(request, cancellationToken);
 		}
 
 		/// <summary>
@@ -103,168 +55,10 @@ namespace Coordix.Implementation
 		/// <param name="notification">The notification message to publish.</param>
 		/// <param name="cancellationToken">A token to monitor for cancellation requests.</param>
 		/// <returns>A task representing the asynchronous publish operation.</returns>
-		public async Task Publish<TNotification>(TNotification notification, CancellationToken cancellationToken = default)
-				where TNotification : INotification
+		public Task Publish<TNotification>(TNotification notification, CancellationToken cancellationToken = default)
+			where TNotification : INotification
 		{
-			var notificationType = notification.GetType();
-			var handlerType = typeof(INotificationHandler<>).MakeGenericType(notificationType);
-			var handlers = _provider.GetServices(handlerType).ToArray();
-			var tasksHandlers = handlers.Where(h => h != null).Select(handler => ToHandleInvokeNotification(handlerType, handler!, notification, cancellationToken));
-
-			await Task.WhenAll(tasksHandlers);
-		}
-
-		/// <summary>
-		/// Gets or creates a cached MethodInfo for the Handle method of a handler type.
-		/// This method ensures that reflection is performed only once per handler type,
-		/// with subsequent calls returning the cached MethodInfo for improved performance.
-		/// </summary>
-		/// <param name="handlerType">The handler type to get the MethodInfo for.</param>
-		/// <returns>The cached MethodInfo for the Handle method.</returns>
-		/// <exception cref="InvalidOperationException">Thrown if the Handle method is not found on the handler type.</exception>
-		private static MethodInfo GetHandleMethodInfo(Type handlerType)
-		{
-			return _methodInfoCache.GetOrAdd(handlerType, type => type.GetMethod("Handle")
-					?? throw new InvalidOperationException($"Handle method not found on {type.Name}"));
-		}
-
-		/// <summary>
-		/// Gets or creates a cached delegate for invoking a request handler without return value.
-		/// </summary>
-		private static Func<object, IRequest, CancellationToken, Task> GetRequestHandlerDelegate(Type handlerType)
-				=> _requestHandlerDelegates.GetOrAdd(handlerType, CreateRequestHandlerDelegate);
-
-		/// <summary>
-		/// Creates a strongly-typed delegate for invoking a request handler without return value.
-		/// Uses Expression Trees to compile a delegate that directly invokes the handler's Handle method,
-		/// avoiding the overhead of MethodInfo.Invoke on every call.
-		/// </summary>
-		/// <param name="handlerType">The handler type to create a delegate for.</param>
-		/// <returns>A compiled delegate that can invoke the handler's Handle method.</returns>
-		private static Func<object, IRequest, CancellationToken, Task> CreateRequestHandlerDelegate(Type handlerType)
-		{
-			var methodInfo = GetHandleMethodInfo(handlerType);
-			var handlerParam = Expression.Parameter(typeof(object), "handler");
-			var requestParam = Expression.Parameter(typeof(IRequest), "request");
-			var cancellationTokenParam = Expression.Parameter(typeof(CancellationToken), "cancellationToken");
-
-			// Extract the concrete request type from IRequestHandler<TRequest>
-			var requestType = handlerType.GetGenericArguments()[0];
-
-			var castHandler = Expression.Convert(handlerParam, handlerType);
-			var castRequest = Expression.Convert(requestParam, requestType);
-			var call = Expression.Call(castHandler, methodInfo, castRequest, cancellationTokenParam);
-			var lambda = Expression.Lambda<Func<object, IRequest, CancellationToken, Task>>(call, handlerParam, requestParam, cancellationTokenParam);
-
-			return lambda.Compile();
-		}
-
-		/// <summary>
-		/// Gets or creates a cached delegate for invoking a request handler with return value.
-		/// </summary>
-		private static Func<object, IRequest<TResponse>, CancellationToken, Task<TResponse>> GetRequestHandlerWithResponseDelegate<TResponse>(Type handlerType)
-				=> (Func<object, IRequest<TResponse>, CancellationToken, Task<TResponse>>)_requestHandlerWithResponseDelegates.GetOrAdd(handlerType, _ => CreateRequestHandlerWithResponseDelegate<TResponse>(handlerType));
-
-		/// <summary>
-		/// Creates a strongly-typed delegate for invoking a request handler with return value.
-		/// Uses Expression Trees to compile a delegate that directly invokes the handler's Handle method,
-		/// avoiding the overhead of MethodInfo.Invoke on every call.
-		/// </summary>
-		/// <typeparam name="TResponse">The type of the response expected from the handler.</typeparam>
-		/// <param name="handlerType">The handler type to create a delegate for.</param>
-		/// <returns>A compiled delegate that can invoke the handler's Handle method and return the response.</returns>
-		private static Func<object, IRequest<TResponse>, CancellationToken, Task<TResponse>> CreateRequestHandlerWithResponseDelegate<TResponse>(Type handlerType)
-		{
-			var methodInfo = GetHandleMethodInfo(handlerType);
-			var handlerParam = Expression.Parameter(typeof(object), "handler");
-			var requestParam = Expression.Parameter(typeof(IRequest<TResponse>), "request");
-			var cancellationTokenParam = Expression.Parameter(typeof(CancellationToken), "cancellationToken");
-
-			// Extract the concrete request type from IRequestHandler<TRequest, TResponse>
-			var requestType = handlerType.GetGenericArguments()[0];
-
-			var castHandler = Expression.Convert(handlerParam, handlerType);
-			var castRequest = Expression.Convert(requestParam, requestType);
-			var call = Expression.Call(castHandler, methodInfo, castRequest, cancellationTokenParam);
-			var lambda = Expression.Lambda<Func<object, IRequest<TResponse>, CancellationToken, Task<TResponse>>>(call, handlerParam, requestParam, cancellationTokenParam);
-
-			return lambda.Compile();
-		}
-
-		/// <summary>
-		/// Gets or creates a cached delegate for invoking a notification handler.
-		/// </summary>
-		private static Func<object, INotification, CancellationToken, Task> GetNotificationHandlerDelegate(Type handlerType)
-				=> _notificationHandlerDelegates.GetOrAdd(handlerType, CreateNotificationHandlerDelegate);
-
-		/// <summary>
-		/// Creates a strongly-typed delegate for invoking a notification handler.
-		/// Uses Expression Trees to compile a delegate that directly invokes the handler's Handle method,
-		/// avoiding the overhead of MethodInfo.Invoke on every call.
-		/// </summary>
-		/// <param name="handlerType">The handler type to create a delegate for.</param>
-		/// <returns>A compiled delegate that can invoke the handler's Handle method.</returns>
-		private static Func<object, INotification, CancellationToken, Task> CreateNotificationHandlerDelegate(Type handlerType)
-		{
-			var methodInfo = GetHandleMethodInfo(handlerType);
-			var handlerParam = Expression.Parameter(typeof(object), "handler");
-			var notificationParam = Expression.Parameter(typeof(INotification), "notification");
-			var cancellationTokenParam = Expression.Parameter(typeof(CancellationToken), "cancellationToken");
-
-			// Extract the concrete notification type from INotificationHandler<TNotification>
-			var notificationType = handlerType.GetGenericArguments()[0];
-
-			var castHandler = Expression.Convert(handlerParam, handlerType);
-			var castNotification = Expression.Convert(notificationParam, notificationType);
-			var call = Expression.Call(castHandler, methodInfo, castNotification, cancellationTokenParam);
-			var lambda = Expression.Lambda<Func<object, INotification, CancellationToken, Task>>(call, handlerParam, notificationParam, cancellationTokenParam);
-
-			return lambda.Compile();
-		}
-
-		/// <summary>
-		/// Invokes the Handle method on a notification handler.
-		/// </summary>
-		/// <typeparam name="TNotification">The notification type.</typeparam>
-		/// <param name="handlerType">The runtime handler interface type.</param>
-		/// <param name="handler">The handler instance to invoke.</param>
-		/// <param name="notification">The notification message.</param>
-		/// <param name="cancellationToken">A token to monitor for cancellation requests.</param>
-		/// <returns>A task representing the handler invocation.</returns>
-		private static Task ToHandleInvokeNotification<TNotification>(Type handlerType, object handler, TNotification notification, CancellationToken cancellationToken)
-				where TNotification : INotification
-		{
-			var delegateFunc = GetNotificationHandlerDelegate(handlerType);
-			return delegateFunc(handler, notification, cancellationToken);
-		}
-
-		/// <summary>
-		/// Invokes the Handle method on a request handler without return value.
-		/// </summary>
-		/// <param name="handlerType">The runtime handler interface type.</param>
-		/// <param name="handler">The handler instance to invoke.</param>
-		/// <param name="request">The request message.</param>
-		/// <param name="cancellationToken">A token to monitor for cancellation requests.</param>
-		/// <returns>A task representing the handler invocation.</returns>
-		private static Task ToHandleInvokeRequest(Type handlerType, object handler, IRequest request, CancellationToken cancellationToken)
-		{
-			var delegateFunc = GetRequestHandlerDelegate(handlerType);
-			return delegateFunc(handler, request, cancellationToken);
-		}
-
-		/// <summary>
-		/// Invokes the Handle method on a request handler and returns its response.
-		/// </summary>
-		/// <typeparam name="TResponse">The type of the response expected from the handler.</typeparam>
-		/// <param name="handlerType">The runtime handler interface type.</param>
-		/// <param name="handler">The handler instance to invoke.</param>
-		/// <param name="request">The request message.</param>
-		/// <param name="cancellationToken">A token to monitor for cancellation requests.</param>
-		/// <returns>A task representing the handler invocation, containing the handler's response.</returns>
-		private static Task<TResponse> ToHandleInvokeRequest<TResponse>(Type handlerType, object handler, IRequest<TResponse> request, CancellationToken cancellationToken)
-		{
-			var delegateFunc = GetRequestHandlerWithResponseDelegate<TResponse>(handlerType);
-			return delegateFunc(handler, request, cancellationToken);
+			return _handlerExecutor.ExecuteNotificationHandler(notification, cancellationToken);
 		}
 	}
 }
